+++ src/components/ArrowIcon/style.tsx
@@ -1,12 +1,11 @@
 import styled from 'styled-components';
 
 const Style = styled.div`
-  border: 1px solid black;
   cursor: pointer;
   display: inline-block;
   position: fixed;
-  right: 25px;
-  top: 25px;
+  right: 22px;
+  top: 22px;
 `;
 
 export default Style;